support_manager/requirements.txt
python-ldap==3.4.3
flask-ldap3-login==0.9.18
flask-wtf==1.0.1
sqlalchemy==1.4.42
flask-sqlalchemy==2.5.1
flask-login==0.6.2
ecs-logging==2.0.0
concurrent-log-handler==0.9.20
pyopenssl==22.1.0
Werkzeug~=2.0.0
requests==2.28.1
